Fix build issue

+++ IdentityCore/src/telemetry/request_telemetry/MSIDLastRequestTelemetrySerializedItem.m
@@ -83,7 +83,7 @@
         // Fill in remaining errors with comma at the end of each error
         for (int i = lastIndex - 1; i >= 0; i--)
         {
-            NSString *currentFailedRequest = [NSString stringWithFormat:@"%ld,%@,", self.errorsInfo[i].apiId, self.errorsInfo[i].correlationId ?: @""];
+            NSString *currentFailedRequest = [NSString stringWithFormat:@"%ld,%@,", (long)self.errorsInfo[i].apiId, self.errorsInfo[i].correlationId ?: @""];
             NSString *currentErrorMessage = [NSString stringWithFormat:@"%@,", self.errorsInfo[i].error ?: @""];
             
             NSString *newFailedRequestsString = [currentFailedRequest stringByAppendingString:failedRequestsString];
